Fix memory corruption caused by swprintf
swprintf requires a destination buffer size like snprintf. This was
previously only working by chance because of how MSVC emits variadic
functions.

swprintf also automatically appends a null terminator to the output
string so there's no reason to add it via the format string.

This was leading to a buffer overflow and occasional heap corruption.